refactor: restore reis seam by moving pure PMTiles helpers to a leaf module
`uv run lint-imports` was failing on main with 2 broken contracts: the
validation-seam-for-reis and validation-no-framework-leakage contracts (added
in #591) were violated by validation.rules -> pmtiles -> output/style ->
click/rich/config. The pmtiles import landed in the pmtiles_link rule via #586,
whose branch post-dated #591's branch, so each PR was green alone and the
conflict only appeared once both merged. import-linter runs as the
"architecture rules" pre-commit hook, so this left main un-committable.

The pmtiles_link rule only needs three framework-free symbols from pmtiles.py:
WEB_MAP_LINKS_EXTENSION, pmtiles_asset_hrefs, pmtiles_link_hrefs (plus the
PMTILES_MEDIA_TYPE they depend on). Extract these into a new stdlib-only leaf
module portolan_cli/pmtiles_links.py and:

- re-import the two constants back into pmtiles.py (both used internally), so
  its public API is unchanged for existing callers/tests
- point validation/rules.py at the leaf (restores the seam)
- split metadata/fix.py's import: pure helpers from the leaf, the impure
  add_pmtiles_link_to_collection/ensure_web_map_links_extension from pmtiles
- document the intentional validation -> pmtiles_links coupling in the
  import-linter contract comment

Also add `lint-imports` to the CI lint job. It was previously only a
pre-commit hook (not in CI), which is why the #586/#591 semantic merge conflict
landed on main un-noticed: each PR was green on its own branch and nothing
re-checked the contracts against the merged tree.

Restores all 5 contracts to KEPT. No behavior change; the helpers are moved
verbatim.
